Basic work on local transport
Summary:
This diff provides the initial groundwork on the local transport. It includes a config manager for the API and client, as well as the basic infrastructure to start up the local transport and communicate with the client over xmlrpc.

The client will continue to run after the API has terminated, and the API will attempt to reconnect to the client on the previous port when it is run next using the local transport. The client will continue to run in the background until `terminate()` is called on the transport.

Test Plan:
The following test script should run fine. The client will report the script to be run in the client log file (`~/.config/finorch/logs/` on linux).

```from finorch.transport.local import LocalTransport

transport = LocalTransport()
transport.connect()

script = """
    test
"""

jid = transport.start_job(script)

print(jid)

# transport.terminate()```

Reviewers: ssaleheen

Reviewed By: ssaleheen

Differential Revision: https://phab.adacs.org.au/D721