Basic Session/Client/Wrapper structure
Summary:
This diff provides the base session/client/wrapper file structure to aid in more quickly adding a new client later on. It improves the termination code of the client's XMLRPC server as well.

In this context a Session should represent the API instance that the user should interact with. For example a LocalSession(), or an OzSTARSession(username="blah"). The Client is the pseudo-daemon that manages finesse jobs. The Client is responsible for communicating with the Wrapper which itself is the wrapper around the finesse process. There may be many Wrapper/finesse processes running, and the client orchestrates these via whatever logic is required for the specific site (local, ozstar, cit etc).

Test Plan: The same script provided in D721 should work the same way. However the client python process is now correctly cleaned up when calling transport.terminate().

Reviewers: ssaleheen, ddbrown

Reviewed By: ssaleheen

Differential Revision: https://phab.adacs.org.au/D722