Initial ability to run Local jobs via katscript
Summary: This diff adds in the local wrapper and modifies the code to a state where it is possible to submit local finesse jobs using a multiprocessing pool.

Test Plan:
Running the following test script should result in a 128 finesse jobs being successfully executed and the output placed in `~/finorch/*uuids*/`. The test script should also print the UUIDs representing the jobs that has been executed.

```
session = LocalSession(exec_path="~/finorch/")

script = """
    # Add a Laser named L0 with a power of 1 W.
    l L0 P=1

    # Space attaching L0 <-> m1 with length of 0 m (default).
    s s0 L0.p1 m1.p1

    # Input mirror of cavity.
    m m1 R=0.99 T=0.01

    # Intra-cavity space with length of 1 m.
    s CAV m1.p2 m2.p1 L=1

    # End mirror of cavity.
    m m2 R=0.991 T=0.009

    # Power detectors on reflection, circulation and transmission.
    pd refl m1.p1.o
    pd circ m2.p1.i
    pd trns m2.p2.o

    # Scan over the detuning DOF of m1 from -180 deg to +180 deg with 400 points.
    xaxis(m1.phi, lin, -180, 180, 400)
"""

for i in range(128):
    job = session.start_job(script)
    print(job)
```

Reviewers: ssaleheen, ddbrown

Reviewed By: ssaleheen

Differential Revision: https://phab.adacs.org.au/D723