Add database management to local jobs
Summary:
To avoid trying to fetch job solution objects for jobs that are not yet complete it's important to track the state of jobs. This diff adds a database class that keeps a record of job status. Jobs which are not complete yet are updated when a request for job status is made, job status for jobs which have a completion status will be read directly from the database. The code to fetch a solution object has been updated to check that the job has finished.

Due to the nature of multiprocessing in python, the simplest way to track if a job is queued, running or finished is to create marker files in each job output. When a client wrapper starts a job, it will create a 'started' file, and upon completion it will create a 'finished' file. These are used to determine in which state the job is in.

Test Plan: When running jobs, local jobs that are queued should raise an exception upon attempting to retrieve the solution object. It should be possible to retrieve the status of any submitted job.

Reviewers: ssaleheen

Reviewed By: ssaleheen

Differential Revision: https://phab.adacs.org.au/D744