Add SSH key management
Summary: This diff introduces the ability for a session using SshTransport to use ssh keys rather than a password. Two helper scripts have been added that allow the management of the ssh keys for various sessions.

Test Plan: After running `poetry install`, it should be possible to use the `set_ssh_key` executable to configure an ssh key to use for the ozstar session. Once the key is configured, you can connect to the ozstar session without providing a password.

Reviewers: ssaleheen

Reviewed By: ssaleheen

Differential Revision: https://phab.adacs.org.au/D762