Implement stop_job for OzSTAR
Summary: This diff adds the functionality required to use the stop_job function on OzSTAR. It also removes the update_parameters function and fixes up various tests.

Test Plan:
If a job is started using the OzSTAR session, and `squeue -u whoeveryouruseris` is executed, slurm should report a queued (or running) job.
Then calling stop_job on that job id should remove the job from the queue or cancel the running job. Calling job_status on that job should then report a status of "CANCELLED".

Reviewers: ssaleheen

Reviewed By: ssaleheen

Differential Revision: https://phab.adacs.org.au/D769